Test best close functaion

+++ K-Means/OmpFunctions.h
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <math.h>
 
-double distance2Points(point_t p1, point_t p2);
+float distance2Points(point_t p1, point_t p2);
 void setClosestCluster(point_t* points, int numberOfPoints, point_t* clusters, int numberOfCluster);
 point_t* averageClusters(point_t* points, int numberOfPoints, int numberOfCluster);
 point_t* combainPointsArrays(point_t* points, int numberOfArrays, int pointsInArray);
